MainProject/pyproject.toml
[tool.poetry]
name = "mainproject"
version = "0.1.0"
description = ""
authors = ["AnanasekE <janeksambor@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"


[tool.poetry.group.dev.dependencies]
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
